refactor: Restructure transit data handling
