--- app/src/main/java/com/will/busnotification/data/dto/places/LegsDTO.kt
@@ -1,7 +1,0 @@
-package com.will.busnotification.data.dto.places
-
-import com.will.busnotification.data.dto.PlaceResult
-
-data class LegsDTO {
-    val steps: List<StepsDTO>
-}
