--- app/src/main/java/com/will/busnotification/data/dto/places/PlacesResponse.kt
@@ -1,5 +1,0 @@
-package com.will.busnotification.data.dto.places
-
-data class PlacesResponse(
-    val routes: List<RoutesDTO>,
-)
