--- app/src/main/java/com/will/busnotification/data/dto/places/RoutesDTO.kt
@@ -1,5 +1,0 @@
-package com.will.busnotification.data.dto.places
-
-data class RoutesDTO {
-    val legs: List<LegsDTO>
-}
