--- app/src/main/java/com/will/busnotification/data/dto/places/StepsDTO.kt
@@ -1,5 +1,0 @@
-package com.will.busnotification.data.dto.places
-
-data class StepsDTO {
-    val transitDetails: TransitDetailsDTO?
-}
