+++ app/src/main/java/com/will/busnotification/data/network/GooglePlacesApiService.kt
@@ -1,6 +1,6 @@
 package com.will.busnotification.data.network
 
-import com.will.busnotification.data.dto.places.PlacesResponse
+import com.will.busnotification.data.dto.DirectionsResponseDto
 import com.will.busnotification.data.dto.RouteRequest
 import retrofit2.http.Body
 import retrofit2.http.Header
@@ -12,7 +12,7 @@ interface GooglePlacesApiService {
     @POST("directions/v2:computeRoutes")
     suspend fun searchPlaces(
         @Header("X-Goog-Api-Key") apiKey: String,
-        @Header("X-Goog-FieldMask") fieldMask: String = "routes.duration,routes.distanceMeters",
+        @Header("X-Goog-FieldMask") fieldMask: String = "routes.legs.steps.transitDetails",
         @Body request: RouteRequest
-    ): PlacesResponse
+    ): DirectionsResponseDto
 }
