+++ app/src/main/java/com/will/busnotification/repository/PlacesRepository.kt
@@ -1,7 +1,7 @@
 package com.will.busnotification.repository
 
-import com.will.busnotification.data.dto.PlaceResult
+import com.will.busnotification.data.model.TransitSegment
 
 interface PlacesRepository {
-    suspend fun searchPlaces(query: String): List<PlaceResult>
+    suspend fun searchPlaces(query: String): List<TransitSegment>
 }
